backend/requirements.txt
# ══════════════════════════════════════════════
#  REMEHBS API — Dépendances Python
# ══════════════════════════════════════════════

# Django
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0
dj-database-url==3.1.2
python-dotenv==1.2.2

# Base de données
psycopg2-binary==2.9.11

# Serveur WSGI
gunicorn==25.3.0
whitenoise==6.12.0

# Celery & Redis
celery==5.6.3
django-celery-beat==2.9.0
redis==6.4.0

# Email
resend==2.29.0

# Utilitaires
requests==2.33.1
reportlab==4.4.10
pillow==12.2.0
sentry-sdk==2.40.0
